backend/requirements.txt
fastapi==0.78.0
PyJWT==2.4.0
sqlalchemy==1.4.39
uvicorn==0.18.2
pydantic==1.9.1
cryptography==37.0.2
passlib==1.7.4
PyMySQL==1.0.2
pytest==7.1.2
docker==5.0.3
bcrypt==3.2.0
python-dotenv==0.20.0
mysql-connector-python==8.0.17